ADD 增加队伍加入人数记录

+++ src/main/java/com/beeran/backend/model/vo/TeamUserVO.java
@@ -68,7 +68,7 @@ public class TeamUserVO implements Serializable {
     /**
      * 已加入的用户数
      */
-    private Integer hasJoinNum;
+    private Long hasJoinNum;
 
     /**
      * 是否已加入队伍
